+++ src/components/detail/comments/ReCommentLikes.tsx
@@ -7,7 +7,7 @@ import styled from 'styled-components';
 import supabase from 'src/lib/supabaseClient';
 import { IconLiked, IconUnLiked } from 'src/components/icons';
 import { useLocation } from 'react-router';
-import { NON_MEMBER } from 'src/utility/message';
+import { NON_MEMBER } from 'src/utility/guide';
 import { styleFont } from 'src/styles/styleFont';
 import { toast } from 'react-toastify';
