+++ src/components/renderPosts/reactionSource/WriterContainer.tsx
@@ -4,7 +4,7 @@ import CreatedAt from 'src/utility/CreatedAt';
 import styled, { css } from 'styled-components';
 import { useNavigate } from 'react-router';
 import usePost from 'src/hooks/usePost';
-import { CANCLE, DELETE } from 'src/utility/message';
+import { CANCLE, DELETE } from 'src/utility/guide';
 import { styleFont } from 'src/styles/styleFont';
 import UserLevel from 'src/components/header/UserLevel';
 import { IconClose } from 'src/components/icons';
